Updated dependencies, dropped old nodes and added sessiontTimeout option
