+++ .travis.yml
@@ -3,4 +3,3 @@ language: node_js
 node_js:
   - "12"
   - "10"
-  - "8"
