+++ lib/request.js
@@ -16,6 +16,7 @@ function buildRequest (opts) {
   }
   const baseUrl = opts.base
   const rejectUnauthorized = opts.rejectUnauthorized
+  const sessionTimeout = opts.sessionTimeout || 60000
   var http2Client
   var pool
   var agents
@@ -97,7 +98,10 @@ function buildRequest (opts) {
       // we might enqueue a large number of requests in this connection
       // before it's connected
       http2Client.setMaxListeners(0)
-      http2Client.on('connect', () => {
+      http2Client.setTimeout(sessionTimeout, function () {
+        http2Client.destroy()
+      })
+      http2Client.once('connect', () => {
         // reset the max listener to 10 on connect
         http2Client.setMaxListeners(10)
         http2Client.removeListener('error', done)
