+++ test/http2-invalid-target.js
@@ -2,7 +2,7 @@
 
 const { test } = require('tap')
 const Fastify = require('fastify')
-const From = require('../..')
+const From = require('..')
 const got = require('got')
 
 test('http -> http2', async (t) => {
@@ -14,7 +14,7 @@ test('http -> http2', async (t) => {
     reply.from()
   })
   instance.register(From, {
-    base: `http://abc.xyz1`,
+    base: 'http://abc.xyz1',
     http2: true
   })
